Rename styles.scss -> OnboardingGuide.css
