+++ client/app/components/OnboardingGuide/GuideStatusPage.js
@@ -2,7 +2,7 @@ import { PropTypes } from 'react';
 import { div, p, h2, hr, ul, li, span, a } from 'r-dom';
 import { t } from '../../utils/i18n';
 
-import css from './styles.scss';
+import css from './OnboardingGuide.css';
 
 const GuideStatusPage = (props) => {
   const handleClick = function handleClick(e, path) {
